feat: Add functionality to request a new token
This commit adds the `request_new_token()` function to the `io.py` file. This function makes an HTTP request to the authentication endpoint and returns a new token if the request is successful. If the request fails, an exception is raised.

Additionally, in the `ably_connect()` function, if there is an error during authentication, a new token is requested using the `request_new_token()` function. The new token is then saved in a token file and another attempt to connect is made.

The `DIFF_IMPLEMENTATION_PLAN.md` file has been deleted as it is no longer needed.

In the `setup.py` file, updated import statements have been added for running the client using different command line scripts.

Closes #123